+++ rt/rest2.py
@@ -409,7 +409,7 @@ class Rt:
 
         return res
 
-    def new_correspondence(self, queue: typing.Optional[typing.Union[str, object]] = None) -> typing.Iterator[dict]:
+    def new_correspondence(self, queue: typing.Optional[typing.Union[str, object]] = None) -> typing.Iterator[dict[str, typing.Any]]:
         """Obtains tickets changed by other users than the system one.
 
         :param queue: Queue where to search
@@ -421,7 +421,7 @@ class Rt:
         """
         return self.search(queue=queue, order='-LastUpdated')
 
-    def last_updated(self, since: str, queue: typing.Optional[str] = None) -> typing.Iterator[dict]:
+    def last_updated(self, since: str, queue: typing.Optional[str] = None) -> typing.Iterator[dict[str, typing.Any]]:
         """Obtains tickets changed after given date.
 
         :param since: Date as string in form '2011-02-24'
@@ -460,9 +460,9 @@ class Rt:
         queue: typing.Optional[typing.Union[str, object]] = None,
         order: typing.Optional[str] = None,
         raw_query: typing.Optional[str] = None,
-        query_format: typing.Union[str, list[str]] = 'l',
+        query_format: typing.Union[str, list[str], dict[str, str]] = 'l',
         **kwargs: typing.Any,
-    ) -> typing.Iterator[dict]:
+    ) -> typing.Iterator[dict[str, typing.Any]]:
         r"""Search arbitrary needles in given fields and queue.
 
         Example::
@@ -554,6 +554,8 @@ class Rt:
 
         if isinstance(query_format, list):
             get_params['fields'] = ','.join(query_format)
+        elif isinstance(query_format, dict):
+            get_params = {**get_params, **query_format}
         elif query_format == 'l':
             get_params['fields'] = (
                 'Owner,Status,Created,Subject,Queue,CustomFields,Requestor,Cc,AdminCc,Started,Created,TimeEstimated,Due,Type,InitialPriority,Priority,TimeLeft,LastUpdated'
@@ -564,13 +566,16 @@ class Rt:
 
         yield from self.__paged_request(url, params=get_params)
 
-    def get_ticket(self, ticket_id: typing.Union[str, int]) -> dict:
+    def get_ticket(
+        self, ticket_id: typing.Union[str, int], query_format: typing.Union[dict[str, str], None] = None
+    ) -> dict[str, typing.Any]:
         """Fetch ticket by its ID.
 
         :param ticket_id: ID of demanded ticket
+        :param query_format: Returned fields to be populated
 
         :returns: Dictionary with key, value pairs for ticket with
-                  *ticket_id* or None if ticket does not exist. List of keys:
+                  *ticket_id* or None if ticket does not exist. List of keys (when query_format is None):
 
                       * id
                       * numerical_id
@@ -597,7 +602,10 @@ class Rt:
         :raises UnexpectedMessageFormatError: Unexpected format of returned message.
         :raises NotFoundError: If there is no ticket with the specified ticket_id.
         """
-        res = self.__request(f'ticket/{ticket_id}', get_params={'fields[Queue]': 'Name'})
+        if not query_format:
+            query_format = {}
+
+        res = self.__request(f'ticket/{ticket_id}', get_params={'fields[Queue]': 'Name', **query_format})
 
         if not isinstance(res, dict):  # pragma: no cover
             raise UnexpectedResponseError(str(res))
@@ -916,7 +924,7 @@ class Rt:
 
         return attachments
 
-    def get_attachment(self, attachment_id: typing.Union[str, int]) -> dict:
+    def get_attachment(self, attachment_id: typing.Union[str, int]) -> dict[str, typing.Any]:
         """Get attachment.
 
         :param attachment_id: ID of attachment to fetch
@@ -1628,12 +1636,13 @@ class Rt:
 
         return response
 
-    def get_asset(self, asset_id: typing.Union[str, int]) -> dict[str, typing.Any]:
+    def get_asset(self, asset_id: typing.Union[str, int], query_format: typing.Optional[dict[str, str]] = None) -> dict[str, typing.Any]:
         """
         Get asset.
 
         :param asset_id: Asset ID.
-        :return: Asset.
+        :param query_format: Returned fields to be populated.
+        :return: Asset (when `query_format` is `None`).
                 id: int
                 Lifecycle: str
                 Disabled: str
@@ -1651,7 +1660,10 @@ class Rt:
                 Owner: dict[str, str]
                 CustomFields: list[dict[str, typing.Any]]
         """
-        response = self.__request(f'asset/{asset_id}')
+        if not query_format:
+            query_format = {}
+
+        response = self.__request(f'asset/{asset_id}', get_params=query_format)
 
         self.logger.debug(str(response))
 
@@ -1696,23 +1708,28 @@ class Rt:
         return isinstance(response, list)
 
     def search_assets(
-        self, catalog_id: typing.Union[str, int], search_params: list[dict[str, typing.Any]], fields: str = "Owner,Description,Status"
+        self,
+        catalog_id: typing.Union[str, int],
+        search_params: list[dict[str, typing.Any]],
+        query_format: typing.Optional[typing.Union[str, list[str], dict[str, str]]] = None,
     ) -> typing.Iterator[dict[str, typing.Any]]:
         """
         Search assets in a catalog.
 
         Example::
 
-            client = Rt(...)
-            client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}])
+            >>> client = Rt(...)
+            >>> client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], "Owner,Status")
+            >>> client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], ["Owner", "Status"])
+            >>> client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], {"fields": "Owner,Status", "fields[Owner]": "id,Name"})
 
         :param catalog_id: Catalog ID.
         :param search_params: Params used to filter the results.
             field: str
             value: str | int
             operator: Literal[">", "<", "=", "!=", "LIKE", "NOT LIKE", ">=", "<="] | None
-        :param fields: Fields to return separated by a comma.
-        :return: Found assets. The following is returned with the default `fields`
+        :param query_format: Returned fields to be populated.
+        :return: Found assets. The following is returned by default (when `query_format` is `None`)
             {
                 'Description': '',
                 'id': '1',
@@ -1724,7 +1741,15 @@ class Rt:
         """
         search_params.append({'field': 'Catalog', 'value': catalog_id, 'operator': '='})
 
-        yield from self.__paged_request('assets', json_data=search_params, params={"fields": fields})
+        get_params = {'fields': 'Owner,Description,Status'}
+        if isinstance(query_format, dict):
+            get_params = {**get_params, **query_format}
+        elif isinstance(query_format, list):
+            get_params['fields'] = ','.join(query_format)
+        elif isinstance(query_format, str):
+            get_params['fields'] = query_format
+
+        yield from self.__paged_request('assets', json_data=search_params, params=get_params)
 
     def get_asset_history(self, asset_id: typing.Union[str, int]) -> typing.Iterator[dict[str, typing.Any]]:
         """
@@ -1961,7 +1986,7 @@ class AsyncRt:
         page: int = 1,
         per_page: int = 20,
         recurse: bool = True,
-    ) -> collections.abc.AsyncIterator:
+    ) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Request using pagination for :term:`API`.
 
         :param selector: End part of URL which completes self.url parameter
@@ -2080,7 +2105,9 @@ class AsyncRt:
 
         return res
 
-    async def new_correspondence(self, queue: typing.Optional[typing.Union[str, object]] = None) -> collections.abc.AsyncIterator:
+    async def new_correspondence(
+        self, queue: typing.Optional[typing.Union[str, object]] = None
+    ) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Obtains tickets changed by other users than the system one.
 
         :param queue: Queue where to search
@@ -2089,11 +2116,11 @@ class AsyncRt:
                   the system one, ordered in decreasing order by LastUpdated.
                   Each ticket is dictionary, the same as in
                   :py:meth:`~Rt.get_ticket`.
-                  collections.abc.AsyncIterator[dict]
+                  collections.abc.AsyncIterator[dict[str, typing.Any]]
         """
         return self.search(queue=queue, order='-LastUpdated')
 
-    async def last_updated(self, since: str, queue: typing.Optional[str] = None) -> collections.abc.AsyncIterator:
+    async def last_updated(self, since: str, queue: typing.Optional[str] = None) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Obtains tickets changed after given date.
 
         :param since: Date as string in form '2011-02-24'
@@ -2103,7 +2130,7 @@ class AsyncRt:
                   *since* ordered in decreasing order by LastUpdated.
                   Each ticket is a dictionary, the same as in
                   :py:meth:`~Rt.get_ticket`.
-                  collections.abc.AsyncIterator[dict]
+                  collections.abc.AsyncIterator[dict[str, typing.Any]]
 
         :raises InvalidUseError: If the specified date is of an unsupported format.
         """
@@ -2133,9 +2160,9 @@ class AsyncRt:
         queue: typing.Optional[typing.Union[str, object]] = None,
         order: typing.Optional[str] = None,
         raw_query: typing.Optional[str] = None,
-        query_format: typing.Union[str, list[str]] = 'l',
+        query_format: typing.Union[str, list[str], dict[str, str]] = 'l',
         **kwargs: typing.Any,
-    ) -> collections.abc.AsyncIterator:
+    ) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         r"""Search arbitrary needles in given fields and queue.
 
         Example::
@@ -2186,7 +2213,7 @@ class AsyncRt:
 
         :returns: Iterator over matching tickets. Each ticket is the same dictionary
                   as in :py:meth:`~Rt.get_ticket`.
-                  collections.abc.AsyncIterator[dict]
+                  collections.abc.AsyncIterator[dict[str, typing.Any]]
         :raises:  UnexpectedMessageFormatError: Unexpected format of returned message.
                   InvalidQueryError: If raw query is malformed
         """
@@ -2228,6 +2255,8 @@ class AsyncRt:
 
         if isinstance(query_format, list):
             get_params['fields'] = ','.join(query_format)
+        elif isinstance(query_format, dict):
+            get_params = {**get_params, **query_format}
         elif query_format == 'l':
             get_params['fields'] = (
                 'Owner,Status,Created,Subject,Queue,CustomFields,Requestor,Cc,AdminCc,Started,Created,TimeEstimated,Due,Type,InitialPriority,Priority,TimeLeft,LastUpdated'
@@ -2239,13 +2268,16 @@ class AsyncRt:
         async for item in self.__paged_request(url, params=get_params):
             yield item
 
-    async def get_ticket(self, ticket_id: typing.Union[str, int]) -> dict:
+    async def get_ticket(
+        self, ticket_id: typing.Union[str, int], query_format: typing.Union[dict[str, str], None] = None
+    ) -> dict[str, typing.Any]:
         """Fetch ticket by its ID.
 
         :param ticket_id: ID of demanded ticket
+        :param query_format: Returned fields to be populated
 
         :returns: Dictionary with key, value pairs for ticket with
-                  *ticket_id* or None if ticket does not exist. List of keys:
+                  *ticket_id* or None if ticket does not exist. List of keys (when query_format is None):
 
                       * id
                       * numerical_id
@@ -2272,7 +2304,10 @@ class AsyncRt:
         :raises UnexpectedMessageFormatError: Unexpected format of returned message.
         :raises NotFoundError: If there is no ticket with the specified ticket_id.
         """
-        res = await self.__request(f'ticket/{ticket_id}', get_params={'fields[Queue]': 'Name'})
+        if not query_format:
+            query_format = {}
+
+        res = await self.__request(f'ticket/{ticket_id}', get_params={'fields[Queue]': 'Name', **query_format})
 
         if not isinstance(res, dict):  # pragma: no cover
             raise UnexpectedResponseError(str(res))
@@ -2374,7 +2409,7 @@ class AsyncRt:
 
         return bool(msg[0])
 
-    async def get_ticket_history(self, ticket_id: typing.Union[str, int]) -> collections.abc.AsyncIterator:
+    async def get_ticket_history(self, ticket_id: typing.Union[str, int]) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Get set of short history items.
 
         :param ticket_id: ID of ticket
@@ -2527,7 +2562,7 @@ class AsyncRt:
         self,
         ticket_id: typing.Union[str, int],
         query_filter: typing.Optional[list[dict[str, str]]] = None,
-    ) -> collections.abc.AsyncIterator:
+    ) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Get attachment list for a given ticket.
 
         Example of a return result:
@@ -2563,7 +2598,7 @@ class AsyncRt:
         self,
         ticket_id: typing.Union[str, int],
         query_filter: typing.Optional[list[dict[str, str]]] = None,
-    ) -> collections.abc.AsyncIterator:
+    ) -> collections.abc.AsyncIterator[int]:
         """Get IDs of attachments for given ticket.
 
         :param ticket_id: ID of
@@ -2581,7 +2616,7 @@ class AsyncRt:
         ):
             yield int(item['id'])
 
-    async def get_attachment(self, attachment_id: typing.Union[str, int]) -> dict:
+    async def get_attachment(self, attachment_id: typing.Union[str, int]) -> dict[str, typing.Any]:
         """Get attachment.
 
         :param attachment_id: ID of attachment to fetch
@@ -2953,7 +2988,7 @@ class AsyncRt:
 
         return res
 
-    async def get_all_queues(self, include_disabled: bool = False) -> collections.abc.AsyncIterator:
+    async def get_all_queues(self, include_disabled: bool = False) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """Return a list of all queues.
 
         Example of a return result:
@@ -3293,12 +3328,15 @@ class AsyncRt:
 
         return response
 
-    async def get_asset(self, asset_id: typing.Union[str, int]) -> dict[str, typing.Any]:
+    async def get_asset(
+        self, asset_id: typing.Union[str, int], query_format: typing.Optional[dict[str, str]] = None
+    ) -> dict[str, typing.Any]:
         """
         Get asset.
 
         :param asset_id: Asset ID.
-        :return: Asset.
+        :param query_format: Returned fields to be populated.
+        :return: Asset (when `query_format` is `None`).
                 id: int
                 Lifecycle: str
                 Disabled: str
@@ -3316,7 +3354,10 @@ class AsyncRt:
                 Owner: dict[str, str]
                 CustomFields: list[dict[str, typing.Any]]
         """
-        response = await self.__request(f'asset/{asset_id}')
+        if not query_format:
+            query_format = {}
+
+        response = await self.__request(f'asset/{asset_id}', get_params=query_format)
 
         self.logger.debug(str(response))
 
@@ -3361,23 +3402,28 @@ class AsyncRt:
         return isinstance(response, list)
 
     async def search_assets(
-        self, catalog_id: typing.Union[str, int], search_params: list[dict[str, typing.Any]], fields: str = "Owner,Description,Status"
+        self,
+        catalog_id: typing.Union[str, int],
+        search_params: list[dict[str, typing.Any]],
+        query_format: typing.Optional[typing.Union[str, list[str], dict[str, str]]] = None,
     ) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """
         Search assets in a catalog.
 
         Example::
 
-            client = AsyncRt(...)
-            await client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}])
+            >>> client = AsyncRt(...)
+            >>> await client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], "Owner,Status")
+            >>> await client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], ["Owner", "Status"])
+            >>> await client.search_assets(1, [{"field": "Name", "value": "NameOfMyAsset"}], {"fields": "Owner,Status", "fields[Owner]": "id,Name"})
 
         :param catalog_id: Catalog ID.
         :param search_params: Params used to filter the results.
             field: str
             value: str | int
             operator: Literal[">", "<", "=", "!=", "LIKE", "NOT LIKE", ">=", "<="] | None
-        :param fields: Fields to return separated by a comma.
-        :return: Found assets. The following is returned with the default `fields`
+        :param query_format: Returned fields to be populated.
+        :return: Found assets. The following is returned by default (when `query_format` is `None`)
             {
                 'Description': '',
                 'id': '1',
@@ -3389,10 +3435,18 @@ class AsyncRt:
         """
         search_params.append({'field': 'Catalog', 'value': catalog_id, 'operator': '='})
 
-        async for item in self.__paged_request('assets', json_data=search_params, params={"fields": fields}):
+        get_params = {'fields': 'Owner,Description,Status'}
+        if isinstance(query_format, dict):
+            get_params = {**get_params, **query_format}
+        elif isinstance(query_format, list):
+            get_params['fields'] = ','.join(query_format)
+        elif isinstance(query_format, str):
+            get_params['fields'] = query_format
+
+        async for item in self.__paged_request('assets', json_data=search_params, params=get_params):
             yield item
 
-    async def get_asset_history(self, asset_id: typing.Union[str, int]) -> collections.abc.AsyncIterator[list[dict[str, typing.Any]]]:
+    async def get_asset_history(self, asset_id: typing.Union[str, int]) -> collections.abc.AsyncIterator[dict[str, typing.Any]]:
         """
         Get asset history.