+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.5.0], 2026-02-04
+### Added
+- Added option to define all query parameters for get/search asset and ticket endpoints
+- Fixed typing of methods that returned a partially unknown dictionary
+
 ## [v3.4.0], 2025-11-21
 ### Added
 - Added functionality for some of the asset endpoints (get, create, edit, search, get history)